refactor: extend `BalGauges` to `BalPoolsGauges`
