+++ bal_addresses/utils.py
@@ -1,6 +1,4 @@
-import json
 from urllib.request import urlopen
-import requests
 
 
 def get_subgraph_url(chain: str, subgraph="core") -> str:
@@ -31,142 +29,3 @@ def get_subgraph_url(chain: str, subgraph="core") -> str:
             if magic_word in str(line):
                 # url is on next line, return it on the next iteration
                 found_magic_word = True
-
-
-def get_pools_with_rate_provider(chain: str) -> dict:
-    """
-    for every chain, query the official balancer subgraph and retrieve pools that meets
-    all three of the following conditions:
-    - have a rate provider different from address(0)
-    - have a liquidity greater than $250k
-    - either:
-      - have a yield fee > 0
-      - be a meta stable pool with swap fee > 0
-      - be a gyro pool
-
-    params:
-    - chain: name of the chain
-
-    returns:
-    dictionary of the format {chain_name: {pool_id: symbol}}
-    """
-    filtered_pools = {}
-    url = get_subgraph_url(chain)
-    query = """{
-        pools(
-            first: 1000,
-            where: {
-                and: [
-                    {
-                        priceRateProviders_: {
-                            address_not: "0x0000000000000000000000000000000000000000"
-                        }
-                    },
-                    {
-                        totalLiquidity_gt: 250000
-                    },
-                    { or: [
-                        { protocolYieldFeeCache_gt: 0 },
-                        { and: [
-                            { swapFee_gt: 0 },
-                            { poolType_contains: "MetaStable" },
-                            { poolTypeVersion: 1 }
-                        ] },
-                        { poolType_contains_nocase: "Gyro" },
-                    ] }
-                ]
-            }
-        ) {
-            id,
-            symbol
-        }
-    }"""
-    r = requests.post(url, json={"query": query})
-    r.raise_for_status()
-    try:
-        for pool in r.json()["data"]["pools"]:
-            filtered_pools[pool["id"]] = pool["symbol"]
-    except KeyError:
-        # no results for this chain
-        pass
-    return filtered_pools
-
-
-def has_alive_preferential_gauge(chain: str, pool_id: str) -> bool:
-    """
-    check if a pool has an alive preferential gauge using a fresh query to the subgraph
-
-    params:
-    - chain: name of the chain
-    - pool_id: id of the pool
-
-    returns:
-    - True if the pool has a preferential gauge which is not killed
-    """
-    url = get_subgraph_url(chain, "gauges")
-    query = f"""{{
-        liquidityGauges(
-            where: {{
-                poolId: "{pool_id}",
-                isKilled: false,
-                isPreferentialGauge: true
-            }}
-        ) {{
-            id
-        }}
-    }}"""
-    r = requests.post(url, json={"query": query})
-    r.raise_for_status()
-    try:
-        result = r.json()["data"]["liquidityGauges"]
-    except KeyError:
-        result = []
-    if len(result) > 0:
-        return True
-    else:
-        print(f"Pool {pool_id} on {chain} has no alive preferential gauge")
-
-
-def build_core_pools(chain: str):
-    """
-    build the core pools dictionary by taking pools from `get_pools_with_rate_provider` and:
-    - check if the pool has an alive preferential gauge
-    - add pools from whitelist
-    - remove pools from blacklist
-
-    params:
-    chain: name of the chain
-
-    returns:
-    dictionary of the format {pool_id: symbol}
-    """
-    core_pools = get_pools_with_rate_provider(chain)
-
-    # make sure the pools have an alive preferential gauge
-    for pool_id in core_pools.copy():
-        if not has_alive_preferential_gauge(chain, pool_id):
-            del core_pools[pool_id]
-
-    # add pools from whitelist
-    with open("config/core_pools_whitelist.json", "r") as f:
-        whitelist = json.load(f)
-    try:
-        for pool, symbol in whitelist[chain].items():
-            if pool not in core_pools:
-                core_pools[pool] = symbol
-    except KeyError:
-        # no results for this chain
-        pass
-
-    # remove pools from blacklist
-    with open("config/core_pools_blacklist.json", "r") as f:
-        blacklist = json.load(f)
-    try:
-        for pool in blacklist[chain]:
-            if pool in core_pools:
-                del core_pools[pool]
-    except KeyError:
-        # no results for this chain
-        pass
-
-    return core_pools
